# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog App API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Author {
  id: string;
  username: string;
  /** @nullable */
  avatar_url: string | null;
}

export interface Post {
  id: string;
  title: string;
  content: string;
  /** @nullable */
  excerpt: string | null;
  /** @nullable */
  cover_image_url: string | null;
  /** @nullable */
  category: string | null;
  author: Author;
  likes_count: number;
  comments_count: number;
  is_liked: boolean;
  published: boolean;
  created_at: string;
  updated_at: string;
}

export interface PostsResponse {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}

export interface CreatePostRequest {
  title: string;
  content: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  cover_image_url?: string | null;
  /** @nullable */
  category?: string | null;
  published?: boolean;
}

export interface UpdatePostRequest {
  title?: string;
  content?: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  cover_image_url?: string | null;
  /** @nullable */
  category?: string | null;
  published?: boolean;
}

export interface LikeResponse {
  liked: boolean;
  likes_count: number;
}

export interface Comment {
  id: string;
  content: string;
  author: Author;
  post_id: string;
  created_at: string;
}

export interface CreateCommentRequest {
  content: string;
}

export interface Profile {
  id: string;
  username: string;
  email: string;
  /** @nullable */
  avatar_url: string | null;
  /** @nullable */
  bio: string | null;
  posts_count: number;
  created_at: string;
}

export interface UpdateProfileRequest {
  username?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatar_url?: string | null;
}

export type GetPostsParams = {
  page?: number;
  limit?: number;
  category?: string;
  search?: string;
};