# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blog App API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: posts
    description: Blog post operations
  - name: comments
    description: Comment operations
  - name: profile
    description: User profile operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get all published blog posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update an existing blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Post deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like or unlike a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /posts/{id}/comments:
    get:
      operationId: getComments
      tags: [comments]
      summary: Get comments for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Comment deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /profile/posts:
    get:
      operationId: getMyPosts
      tags: [profile]
      summary: Get all posts by the current user
      responses:
        "200":
          description: User's posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Author:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar_url:
          type: ["string", "null"]
      required:
        - id
        - username
        - avatar_url

    Post:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        cover_image_url:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        author:
          $ref: "#/components/schemas/Author"
        likes_count:
          type: integer
        comments_count:
          type: integer
        is_liked:
          type: boolean
        published:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - title
        - content
        - excerpt
        - cover_image_url
        - category
        - author
        - likes_count
        - comments_count
        - is_liked
        - published
        - created_at
        - updated_at

    PostsResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - posts
        - total
        - page
        - limit

    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        cover_image_url:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        published:
          type: boolean
      required:
        - title
        - content

    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        cover_image_url:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        published:
          type: boolean

    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likes_count:
          type: integer
      required:
        - liked
        - likes_count

    Comment:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        author:
          $ref: "#/components/schemas/Author"
        post_id:
          type: string
        created_at:
          type: string
      required:
        - id
        - content
        - author
        - post_id
        - created_at

    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    Profile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatar_url:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        posts_count:
          type: integer
        created_at:
          type: string
      required:
        - id
        - username
        - email
        - avatar_url
        - bio
        - posts_count
        - created_at

    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
        bio:
          type: ["string", "null"]
        avatar_url:
          type: ["string", "null"]