# lib/api-zod/src/generated/types/author.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Author {
  id: string;
  username: string;
  /** @nullable */
  avatar_url: string | null;
}