# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Author } from "./author";

export interface Comment {
  id: string;
  content: string;
  author: Author;
  post_id: string;
  created_at: string;
}