# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Author } from "./author";

export interface Post {
  id: string;
  title: string;
  content: string;
  /** @nullable */
  excerpt: string | null;
  /** @nullable */
  cover_image_url: string | null;
  /** @nullable */
  category: string | null;
  author: Author;
  likes_count: number;
  comments_count: number;
  is_liked: boolean;
  published: boolean;
  created_at: string;
  updated_at: string;
}