# lib/api-zod/src/generated/types/postsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from "./post";

export interface PostsResponse {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}