# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  username: string;
  email: string;
  /** @nullable */
  avatar_url: string | null;
  /** @nullable */
  bio: string | null;
  posts_count: number;
  created_at: string;
}