# lib/api-zod/src/generated/types/updatePostRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePostRequest {
  title?: string;
  content?: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  cover_image_url?: string | null;
  /** @nullable */
  category?: string | null;
  published?: boolean;
}