# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blog App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileRequest {
  username?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatar_url?: string | null;
}